+++ previous-versions/sync-for-expenses-version-1/RELEASES.md
@@ -39,3 +39,13 @@ Based on:
 - [go v0.26.0] previous-versions/sync-for-expenses-version-1
 ### Releases
 - [Go v0.26.0] https://github.com/codatio/client-sdk-go/releases/tag/previous-versions/sync-for-expenses-version-1/v0.26.0 - previous-versions/sync-for-expenses-version-1
+
+## 2023-10-30 01:25:48
+### Changes
+Based on:
+- OpenAPI Doc prealpha https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Sync-Expenses-v1.yaml
+- Speakeasy CLI 1.109.0 (2.173.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [go v0.26.1] previous-versions/sync-for-expenses-version-1
+### Releases
+- [Go v0.26.1] https://github.com/codatio/client-sdk-go/releases/tag/previous-versions/sync-for-expenses-version-1/v0.26.1 - previous-versions/sync-for-expenses-version-1